update: 'document.querySelector' -> custom querySelector
