+++ js/component/menu.js
@@ -2,11 +2,12 @@ import { changeCSS } from "../common.js"
 import { 
     menuLogAddTemplate, menuLogDeleteTemplate, menuLogMoveTemplate, menuLogUpdateTemplate
 } from "../templates/template.js";
+import { querySelector } from "../devUtils/querySelector.js";
 
-const menuBar = document.querySelector("#menu");
-const menuOpenBtn = document.querySelector("#menu-open-btn");
-const menuCloseBtn = document.querySelector("#menu-close-btn");
-const menuContent = document.getElementById("menu-content")
+const menuBar = querySelector("#menu");
+const menuOpenBtn = querySelector("#menu-open-btn");
+const menuCloseBtn = querySelector("#menu-close-btn");
+const menuContent = querySelector("#menu-content")
 
 // menu toggle 이벤트 추가
 menuOpenBtn.addEventListener("click", () => { changeCSS(menuBar, "right", 0) })
